+++ src/keycloak/connection.py
@@ -37,15 +37,32 @@ from .exceptions import KeycloakConnectionError
 class ConnectionManager(object):
     """Represents a simple server connection.
 
-    :param base_url: (str) The server URL.
-    :param headers: (dict) The header parameters of the requests to the server.
-    :param timeout: (int) Timeout to use for requests to the server.
-    :param verify: (bool) Verify server SSL.
-    :param proxies: (dict) The proxies servers requests is sent by.
+    :param base_url: The server URL.
+    :type base_url: str
+    :param headers: The header parameters of the requests to the server.
+    :type headers: dict
+    :param timeout: Timeout to use for requests to the server.
+    :type timeout: int
+    :param verify: Verify server SSL.
+    :type verify: bool
+    :param proxies: The proxies servers requests is sent by.
+    :type proxies: dict
     """
 
     def __init__(self, base_url, headers={}, timeout=60, verify=True, proxies=None):
-        """Init method."""
+        """Init method.
+
+        :param base_url: The server URL.
+        :type base_url: str
+        :param headers: The header parameters of the requests to the server.
+        :type headers: dict
+        :param timeout: Timeout to use for requests to the server.
+        :type timeout: int
+        :param verify: Verify server SSL.
+        :type verify: bool
+        :param proxies: The proxies servers requests is sent by.
+        :type proxies: dict
+        """
         self.base_url = base_url
         self.headers = headers
         self.timeout = timeout
@@ -73,7 +90,11 @@ class ConnectionManager(object):
 
     @property
     def base_url(self):
-        """Return base url in use for requests to the server."""
+        """Return base url in use for requests to the server.
+
+        :returns: Base URL
+        :rtype: str
+        """
         return self._base_url
 
     @base_url.setter
@@ -82,7 +103,11 @@ class ConnectionManager(object):
 
     @property
     def timeout(self):
-        """Return timeout in use for request to the server."""
+        """Return timeout in use for request to the server.
+
+        :returns: Timeout
+        :rtype: int
+        """
         return self._timeout
 
     @timeout.setter
@@ -91,7 +116,11 @@ class ConnectionManager(object):
 
     @property
     def verify(self):
-        """Return verify in use for request to the server."""
+        """Return verify in use for request to the server.
+
+        :returns: Verify indicator
+        :rtype: bool
+        """
         return self._verify
 
     @verify.setter
@@ -100,7 +129,11 @@ class ConnectionManager(object):
 
     @property
     def headers(self):
-        """Return header request to the server."""
+        """Return header request to the server.
+
+        :returns: Request headers
+        :rtype: dict
+        """
         return self._headers
 
     @headers.setter
@@ -110,8 +143,10 @@ class ConnectionManager(object):
     def param_headers(self, key):
         """Return a specific header parameter.
 
-        :param key: (str) Header parameters key.
+        :param key: Header parameters key.
+        :type key: str
         :returns: If the header parameters exist, return its value.
+        :rtype: str
         """
         return self.headers.get(key)
 
@@ -122,32 +157,41 @@ class ConnectionManager(object):
     def exist_param_headers(self, key):
         """Check if the parameter exists in the header.
 
-        :param key: (str) Header parameters key.
+        :param key: Header parameters key.
+        :type key: str
         :returns: If the header parameters exist, return True.
+        :rtype: bool
         """
         return self.param_headers(key) is not None
 
     def add_param_headers(self, key, value):
         """Add a single parameter inside the header.
 
-        :param key: (str) Header parameters key.
-        :param value: (str) Value to be added.
+        :param key: Header parameters key.
+        :type key: str
+        :param value: Value to be added.
+        :type value: str
         """
         self.headers[key] = value
 
     def del_param_headers(self, key):
         """Remove a specific parameter.
 
-        :param key: (str) Key of the header parameters.
+        :param key: Key of the header parameters.
+        :type key: str
         """
         self.headers.pop(key, None)
 
     def raw_get(self, path, **kwargs):
         """Submit get request to the path.
 
-        :param path: (str) Path for request.
+        :param path: Path for request.
+        :type path: str
+        :param kwargs: Additional arguments
+        :type kwargs: dict
         :returns: Response the request.
-        :raises: HttpError Can't connect to server.
+        :rtype: Response
+        :raises KeycloakConnectionError: HttpError Can't connect to server.
         """
         try:
             return self._s.get(
@@ -163,10 +207,15 @@ class ConnectionManager(object):
     def raw_post(self, path, data, **kwargs):
         """Submit post request to the path.
 
-        :param path: (str) Path for request.
-        :param data: (dict) Payload for request.
+        :param path: Path for request.
+        :type path: str
+        :param data: Payload for request.
+        :type data: dict
+        :param kwargs: Additional arguments
+        :type kwargs: dict
         :returns: Response the request.
-        :raises: HttpError Can't connect to server.
+        :rtype: Response
+        :raises KeycloakConnectionError: HttpError Can't connect to server.
         """
         try:
             return self._s.post(
@@ -183,10 +232,15 @@ class ConnectionManager(object):
     def raw_put(self, path, data, **kwargs):
         """Submit put request to the path.
 
-        :param path: (str) Path for request.
-        :param data: (dict) Payload for request.
+        :param path: Path for request.
+        :type path: str
+        :param data: Payload for request.
+        :type data: dict
+        :param kwargs: Additional arguments
+        :type kwargs: dict
         :returns: Response the request.
-        :raises: HttpError Can't connect to server.
+        :rtype: Response
+        :raises KeycloakConnectionError: HttpError Can't connect to server.
         """
         try:
             return self._s.put(
@@ -200,19 +254,24 @@ class ConnectionManager(object):
         except Exception as e:
             raise KeycloakConnectionError("Can't connect to server (%s)" % e)
 
-    def raw_delete(self, path, data={}, **kwargs):
+    def raw_delete(self, path, data=None, **kwargs):
         """Submit delete request to the path.
 
-        :param path: (str) Path for request.
-        :param data: (dict) Payload for request.
+        :param path: Path for request.
+        :type path: str
+        :param data: Payload for request.
+        :type data: dict | None
+        :param kwargs: Additional arguments
+        :type kwargs: dict
         :returns: Response the request.
-        :raises: HttpError Can't connect to server.
+        :rtype: Response
+        :raises KeycloakConnectionError: HttpError Can't connect to server.
         """
         try:
             return self._s.delete(
                 urljoin(self.base_url, path),
                 params=kwargs,
-                data=data,
+                data=data or dict(),
                 headers=self.headers,
                 timeout=self.timeout,
                 verify=self.verify,